user_service and comment_service done

+++ comment_service/src/main/java/ba/unsa/etf/nwt/comment_service/controller/CommentController.java
@@ -1,5 +1,7 @@
 package ba.unsa.etf.nwt.comment_service.controller;
 
+import ba.unsa.etf.nwt.comment_service.exception.ResourceNotFoundException;
+import ba.unsa.etf.nwt.comment_service.exception.WrongInputException;
 import ba.unsa.etf.nwt.comment_service.model.Comment;
 import ba.unsa.etf.nwt.comment_service.response.ResponseMessage;
 import ba.unsa.etf.nwt.comment_service.security.CurrentUser;
@@ -8,6 +10,7 @@ import ba.unsa.etf.nwt.comment_service.service.CommentService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
+import javax.annotation.security.RolesAllowed;
 import javax.validation.Valid;
 import java.util.List;
 
@@ -21,9 +24,10 @@ public class CommentController {
         return commentService.getComment();
     }
 
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
     @PostMapping("/comment/{mainRoleId}")
     public ResponseMessage addComment(@Valid @RequestBody Comment comment, @PathVariable Long mainRoleId, @CurrentUser UserPrincipal currentUser) {
-            return commentService.addComment(comment, mainRoleId, currentUser);
+        return commentService.addComment(comment, mainRoleId, currentUser);
     }
 
     @GetMapping("/comment/{commentID}")
@@ -36,18 +40,39 @@ public class CommentController {
         return commentService.getUserComments(username);
     }
 
+    //pregled svih komentara za jednog korisnika
     @GetMapping("/comment/category/{roleType}/{categoryID}")
     public List<Comment> getCategoryComment(@PathVariable Long roleType, @PathVariable Long categoryID){
         return commentService.getCategoryComment(roleType, categoryID);
     }
 
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
     @PutMapping("/comment/{commentID}")
-    public ResponseMessage updateComment(@Valid @RequestBody Comment comment, @PathVariable Long commentID) {
+    public ResponseMessage updateComment(@Valid @RequestBody Comment comment, @PathVariable Long commentID, @CurrentUser UserPrincipal currentUser) {
+
+        Comment newComment = commentService.findById(commentID)
+                .orElseThrow(() -> new ResourceNotFoundException("Comment not found!"));
+
+        //korisnici mogu updateovat samo vlastite komentare
+        if(!currentUser.getUsername().equals(newComment.getUsername())){
+            throw new WrongInputException("This comment doesn't belong to current user!");
+        }
+
         return commentService.updateComment(comment, commentID);
     }
 
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
     @DeleteMapping("/comment/{commentID}")
-    public ResponseMessage deleteComment(@PathVariable Long commentID){
+    public ResponseMessage deleteComment(@PathVariable Long commentID, @CurrentUser UserPrincipal currentUser){
+
+        Comment comment = commentService.findById(commentID)
+                .orElseThrow(() -> new ResourceNotFoundException("Comment not found!"));
+
+        //korisnici mogu obrisati samo vlastite komentare
+        if(!currentUser.getUsername().equals(comment.getUsername())){
+            throw new WrongInputException("This comment doesn't belong to current user!");
+        }
+
         return commentService.deleteComment(commentID);
     }
 }
